Rename entities to entity
