+++ src/main/java/com/compfest/sea/entity/user/model/Role.java
@@ -1,4 +1,4 @@
-package com.compfest.sea.entities.user;
+package com.compfest.sea.entity.user.model;
 
 public enum Role {
     ADMIN, CUSTOMER, MERCHANT;
