+++ src/main/java/com/compfest/sea/usecase/user/UserUsecaseImpl.java
@@ -1,7 +1,7 @@
 package com.compfest.sea.usecase.user;
 
 import java.util.List;
-import com.compfest.sea.entities.user.User;
+import com.compfest.sea.entity.user.model.User;
 import com.compfest.sea.repository.user.UserDAO;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
